+++ Release_Notes.md
@@ -20,6 +20,7 @@
 - AbstractDBInteraction now logs exceptions to the log file
 - Moved check for ConfigSettings.CentralManagement from the BusinessLogic project to the Web.Support project, this reduces the number of times the business logic class is created resulting in better performance
 - Implemented Edit DB Information (Useful for the purpose of testing)
+- Added copy text to clipboard in GUID Helper
 
 ### Bug Fixes
 - Fixed the namespace in the WebApi project controllers from GrowthWare.Web.Angular.Controllers to GrowthWare.Web.Api.Controllers